recortes-api/requirements.txt
backpack==0.1
blinker==1.4
cleo==0.6.8
Django==2.2.6
django-cors-headers==3.1.1
django-filter==2.2.0
djangorestframework==3.10.3
Faker==0.8.18
gunicorn==19.9.0
inflection==0.3.1
lazy-object-proxy==1.4.2
Markdown==3.1.1
orator==0.9.9
pastel==0.1.1
pendulum==1.5.1
psycopg2-binary==2.8.3
pyaml==16.12.2
Pygments==2.4.2
pylev==1.3.0
python-dateutil==2.8.0
pytz==2019.3
pytzdata==2019.3
PyYAML==5.1.2
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
tzlocal==1.5.1
whitenoise==4.1.4
wrapt==1.11.2
